refactor: change structure and cleanup
