+++ src/app/App.tsx
@@ -1,8 +1,8 @@
 import { Provider } from 'react-redux';
 import { store } from './store';
-import { CompaniesTable } from '../widget/CompaniesTable';
 import './styles/normalize.scss';
 import './styles/common.scss';
+import { CompaniesTable } from '../widget/CompaniesTable';
 
 function App() {
   return (
