+++ src/widget/CompaniesTable/index.tsx
@@ -1,3 +1,3 @@
-import { CompaniesTable } from './ui';
+import { CompaniesTable } from './ui/CompaniesTable';
 
 export { CompaniesTable };
