+++ src/widget/CompaniesTable/ui/CompaniesTable/CompaniesTable.module.scss
