+++ src/widget/CompaniesTable/ui/index.ts
@@ -1,0 +1,3 @@
+import { CompaniesTable } from './CompaniesTable';
+
+export { CompaniesTable };
